fix(iteration): analyzer positions+industry+weak-batches+type fixes
- _normalize_summary: handle real horizons+n_trading_days format (IR was all 0)
- _load_positions: new method; _pick_industry_source prefers positions.parquet
- _identify_weak_batches: bottom 20pct by date; relabel in to_markdown
- optimizer: float() cast on weight new_value (was np.float64)
- SimDiagnosis __getattr__ aliases: n_trading_days/overall_ir_3m/overall_win_rate_3m
- 935 passed, no regression

+++ quantmind/iteration/optimizer.py
@@ -236,8 +236,8 @@ class ParameterOptimizer:
 
         suggestions: list[ParameterSuggestion] = []
         for k in self.WEIGHT_KEYS:
-            old = round(cur_weights.get(k, 0.25), 4)
-            new = round(new_weights[k], 4)
+            old = float(round(float(cur_weights.get(k, 0.25)), 4))
+            new = float(round(float(new_weights[k]), 4))
             if abs(new - old) < 0.005:
                 continue   # 变化太小，跳过
             ic_val = ic_vals.get(k, float("nan"))
